+++ src/internal/triangle.c
@@ -52,7 +52,7 @@ void sliTriangleOutline(Mat4 *modelview, Vec4 *color)
 	// prepare our shader object
 	shaderBind(sliBasicShader);
 	shaderUniformMatrix4fv(sliBasicShader, "u_Modelview", 1, (float*)modelview);
-	shaderUniform4f(sliBasicShader, "u_Color", color -> x, color -> y, color -> z, color -> w);
+	shaderUniform4f(sliBasicShader, "u_Color", (float)color -> x, (float)color -> y, (float)color -> z, (float)color -> w);
 
 	// bind appropriate object state and render the object
 	#ifndef USE_GLES
@@ -70,7 +70,7 @@ void sliTriangleFill(Mat4 *modelview, Vec4 *color)
 	// prepare our shader object
 	shaderBind(sliBasicShader);
 	shaderUniformMatrix4fv(sliBasicShader, "u_Modelview", 1, (float*)modelview);
-	shaderUniform4f(sliBasicShader, "u_Color", color -> x, color -> y, color -> z, color -> w);
+	shaderUniform4f(sliBasicShader, "u_Color", (float)color -> x, (float)color -> y, (float)color -> z, (float)color -> w);
 
 	// bind appropriate object state and render the object
 	#ifndef USE_GLES
